+++ internal/database/gormQuery/mappingQuery.go
@@ -9,7 +9,7 @@ import (
 	"gorm.io/gorm"
 )
 
-func CreateOrUpdateMapping(gq *GormQuery, mapping *models.Mapping, checkFunc func(mapping *models.Mapping, project *models.Project) ([]uint32, error), deleteMissingElements bool, update bool) error {
+func CreateOrUpdateMapping(gq *GormQuery, mapping *models.Mapping, checkFunc func(mapping *models.Mapping, project *models.Project) ([]int32, error), deleteMissingElements bool, update bool) error {
 	err := gq.Database.Transaction(func(tx *gorm.DB) error {
 		project := models.Project{
 			Model: models.Model{
@@ -81,7 +81,7 @@ func CreateOrUpdateMapping(gq *GormQuery, mapping *models.Mapping, checkFunc fun
 	return err
 }
 
-func (gq *GormQuery) GetAllMappingsQuery(mappings *[]models.Mapping, projectId int, pageSize int, offset int, sortBy string, sortOrder string) error {
+func (gq *GormQuery) GetAllMappingsQuery(mappings *[]models.Mapping, projectId int32, pageSize int, offset int, sortBy string, sortOrder string) error {
 	err := gq.Database.Transaction(func(tx *gorm.DB) error {
 		if err := tx.Where("project_id = ?", projectId).Preload("Elements.Concept.CodeSystem").Order(fmt.Sprintf("%s %s", sortBy, sortOrder)).Offset(offset).Limit(pageSize).Find(&mappings).Error; err != nil {
 			return err
@@ -101,11 +101,11 @@ func (gq *GormQuery) GetAllMappingsQuery(mappings *[]models.Mapping, projectId i
 
 }
 
-func (gq *GormQuery) CreateMappingQuery(mapping *models.Mapping, checkFunc func(mapping *models.Mapping, project *models.Project) ([]uint32, error)) error {
+func (gq *GormQuery) CreateMappingQuery(mapping *models.Mapping, checkFunc func(mapping *models.Mapping, project *models.Project) ([]int32, error)) error {
 	return CreateOrUpdateMapping(gq, mapping, checkFunc, false, false)
 }
 
-func (gq *GormQuery) GetMappingQuery(mapping *models.Mapping, projectId int, mappingId int64) error {
+func (gq *GormQuery) GetMappingQuery(mapping *models.Mapping, projectId int32, mappingId int32) error {
 	err := gq.Database.Transaction(func(tx *gorm.DB) error {
 		if err := tx.Where("project_id = ?", projectId).Preload("Elements.Concept.CodeSystem").First(mapping, mappingId).Error; err != nil {
 			switch {
@@ -128,7 +128,7 @@ func (gq *GormQuery) GetMappingQuery(mapping *models.Mapping, projectId int, map
 	return err
 }
 
-func (gq *GormQuery) UpdateMappingQuery(mapping *models.Mapping, checkFunc func(mapping *models.Mapping, project *models.Project) ([]uint32, error), deleteMissingElements bool) error {
+func (gq *GormQuery) UpdateMappingQuery(mapping *models.Mapping, checkFunc func(mapping *models.Mapping, project *models.Project) ([]int32, error), deleteMissingElements bool) error {
 	// TODO it has to be checked if the mapping exists in the project. If there is another project which has the same CodeSystem Roles and the projectId of the other project is set in the update mapping url, the mapping would get moved to the other project
 	return CreateOrUpdateMapping(gq, mapping, checkFunc, deleteMissingElements, true)
 }
